refactor(io): extract atomic write primitive
Move atomic file replacement out of the cache module so shared disk writes are categorized under the io domain instead of being coupled to result-cache helpers. This keeps cache.ts focused on cache identity and cache entry operations while preserving the hardened write path for renderer, remote downloads, and attachment output.
