+++ src/core/io/cache.ts
@@ -3,19 +3,17 @@ import {
   chmodSync,
   closeSync,
   existsSync,
-  constants as fsConstants,
   lstatSync,
   mkdirSync,
   openSync,
   readFileSync,
   readSync,
-  renameSync,
   rmSync,
   type Stats,
-  writeSync,
 } from 'node:fs';
 import { tmpdir } from 'node:os';
 import { join } from 'node:path';
+import { atomicWrite } from './atomicWrite.js';
 
 /**
  * Resolve the cache root afresh on every call so a test that sets
@@ -33,7 +31,6 @@ export function getCacheRoot(): string {
 // (especially relevant on shared Linux hosts with /tmp/pdfvision/...) cannot
 // read them.
 const DIR_MODE = 0o700;
-const FILE_MODE = 0o600;
 
 // Reject any cache key that isn't a single safe path segment.
 // Cache callers always supply a hash-derived key, so the safe set is narrow.
@@ -185,44 +182,6 @@ export function setCache(cacheDir: string, key: string, data: string): void {
 // see either the previous version or the fully-written new version, never
 // a partially-written file. O_NOFOLLOW + O_EXCL on the temp path prevents
 // the same symlink-redirect attack as the previous direct-write path.
-export function atomicWrite(finalPath: string, data: Buffer): void {
-  const tmpPath = `${finalPath}.${process.pid}.${Math.random().toString(36).slice(2, 10)}.tmp`;
-  const flags =
-    fsConstants.O_WRONLY | fsConstants.O_CREAT | fsConstants.O_EXCL | (isPosix ? fsConstants.O_NOFOLLOW : 0);
-
-  let fd: number;
-  try {
-    fd = openSync(tmpPath, flags, FILE_MODE);
-  } catch (error) {
-    if (isPosix && error instanceof Error && (error as NodeJS.ErrnoException).code === 'ELOOP') {
-      throw new Error(`Refusing to write at ${tmpPath}: path is a symlink`);
-    }
-    throw error;
-  }
-
-  try {
-    // writeSync may return short on partial writes (large buffers, signals).
-    // Loop until the full buffer has been flushed before renaming.
-    let offset = 0;
-    while (offset < data.length) {
-      offset += writeSync(fd, data, offset, data.length - offset);
-    }
-    if (isPosix) chmodSync(tmpPath, FILE_MODE);
-  } catch (error) {
-    closeSync(fd);
-    rmSync(tmpPath, { force: true });
-    throw error;
-  }
-  closeSync(fd);
-
-  try {
-    renameSync(tmpPath, finalPath);
-  } catch (error) {
-    rmSync(tmpPath, { force: true });
-    throw error;
-  }
-}
-
 export function dropCached(cacheDir: string, key: string): void {
   assertSafeKey(key);
   rmSync(join(cacheDir, key), { force: true });
